API Reference

Types

SciMLLogging.AbstractVerbositySpecifier — Type

AbstractVerbositySpecifier{T} Base for types which specify which log messages are emitted at what level.

SciMLLogging.Verbosity — Module

Verbosity

A module defining verbosity levels and presets for SciMLLogging.

Log Levels

Silent: No output

Info: Informational messages

Warn: Warning messages

Error: Error messages

Level(n): Custom log level with integer value n

Verbosity Presets

None: Minimal verbosity preset

All: Maximum verbosity preset

Minimal: Basic verbosity preset

Standard: Standard verbosity preset

Detailed: Detailed verbosity preset

Macros

SciMLLogging.@SciMLMessage — Macro

A macro that emits a log message based on the log level specified in the option of the AbstractVerbositySpecifier supplied.

f_or_message may be a message String, or a 0-argument function that returns a String.

Usage

To emit a simple string, @SciMLMessage("message", verbosity, :option) will emit a log message with the LogLevel specified in verbosity for the given option.

@SciMLMessage can also be used to emit a log message coming from the evaluation of a 0-argument function. This function is resolved in the environment of the macro call. Therefore it can use variables from the surrounding environment. This may be useful if the log message writer wishes to carry out some calculations using existing variables and use them in the log message.

# String message
@SciMLMessage("Hello", verbose, :test1)

# Function for lazy evaluation
x = 10
y = 20

@SciMLMessage(verbosity, :option) do
    z = x + y
    "Sum: $z"
end

Functions

SciMLLogging.verbosity_to_int — Function

    `verbosity_to_int(verb::Verbosity.LogLevel)`
Takes a `Verbosity.LogLevel` and gives a corresponding integer value.
Verbosity settings that use integers or enums that hold integers are relatively common.
This provides an interface so that these packages can be used with SciMLVerbosity. Each of the basic verbosity levels
are mapped to an integer.

- Silent() => 0
- Info() => 1
- Warn() => 2
- Error() => 3
- Level(i) => i

SciMLLogging.verbosity_to_bool — Function

    `verbosity_to_bool(verb::Verbosity.LogLevel)`
Takes a `Verbosity.LogLevel` and gives a corresponding boolean value.
Verbosity settings that use booleans are relatively common.
This provides an interface so that these packages can be used with SciMLVerbosity.
If the verbosity is `Verbosity.Silent`, then `false` is returned. Otherwise, `true` is returned.

SciMLLogging.SciMLLogger — Function

SciMLLogger(; kwargs...)

Create a logger that routes messages to REPL and/or files based on log level.

Keyword Arguments

info_repl = true: Show info messages in REPL

warn_repl = true: Show warnings in REPL

error_repl = true: Show errors in REPL

info_file = nothing: File path for info messages

warn_file = nothing: File path for warnings

error_file = nothing: File path for errors

SciMLLogging.set_logging_backend — Function

set_logging_backend(backend::String)

Set the logging backend preference. Valid options are:

"logging": Use Julia's standard Logging system (default)

"core": Use Core.println for simple output

Note: You must restart Julia for this preference change to take effect.

SciMLLogging.get_logging_backend — Function

get_logging_backend() -> String

Get the current logging backend preference.

Index

SciMLLogging.Verbosity

SciMLLogging.AbstractVerbositySpecifier

SciMLLogging.SciMLLogger

SciMLLogging.get_logging_backend

SciMLLogging.set_logging_backend

SciMLLogging.verbosity_to_bool

SciMLLogging.verbosity_to_int

SciMLLogging.@SciMLMessage